refactor: consolidate rendering and panel toggles
